refactor: move the project component to a separate folder.
