+++ src/app/[locale]/components/project/ProjectCard.tsx
